requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
async-property==0.2.2
async-timeout==4.0.3
attrs==23.1.0
click==8.1.7
Django==5.0
djangorestframework==3.14.0
drfasyncview==0.2.3
et-xmlfile==1.1.0
exceptiongroup==1.2.0
frozenlist==1.4.0
h11==0.14.0
idna==3.6
install==1.3.5
multidict==6.0.4
openpyxl==3.1.2
pydantic==2.5.2
pydantic_core==2.14.5
python-multipart==0.0.6
pytz==2023.3.post1
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.27.0
typing_extensions==4.8.0
yarl==1.9.4
